fix executable scripts, and rename script
