+++ scripts/legacy/scil_assign_custom_color_to_tractogram.py
