+++ scripts/legacy/scil_assign_uniform_color_to_tractograms.py
