+++ scripts/legacy/scil_compute_bundle_mean_std.py
