+++ scripts/legacy/scil_compute_bundle_mean_std_per_point.py
