+++ scripts/legacy/scil_compute_centroid.py
