+++ scripts/legacy/scil_compute_endpoints_map.py
