+++ scripts/legacy/scil_compute_hdf5_average_density_map.py
