+++ scripts/legacy/scil_compute_mean_fixel_afd_from_bundles.py
