+++ scripts/legacy/scil_compute_mean_fixel_afd_from_hdf5.py
