+++ scripts/legacy/scil_compute_seed_by_labels.py
