+++ scripts/legacy/scil_compute_seed_density_map.py
