+++ scripts/legacy/scil_compute_streamlines_density_map.py
