+++ scripts/legacy/scil_convert_rgb.py
