+++ scripts/legacy/scil_decompose_connectivity.py
