+++ scripts/legacy/scil_estimate_bundles_diameter.py
