+++ scripts/legacy/scil_generate_gradient_sampling.py
