+++ scripts/legacy/scil_merge_sh.py
