+++ scripts/legacy/scil_print_header.py
