+++ scripts/legacy/scil_save_connections_from_hdf5.py
