+++ scripts/legacy/scil_score_bundles.py
