+++ scripts/legacy/scil_score_tractogram.py
