+++ scripts/legacy/scil_uniformize_streamlines_endpoints.py
@@ -2,11 +2,11 @@
 # -*- coding: utf-8 -*-
 
 from scilpy.io.deprecator import deprecate_script
-from scripts.scil_tractogram_uniformize_endpoints import main as new_main
+from scripts.scil_bundle_uniformize_endpoints import main as new_main
 
 
 DEPRECATION_MSG = """
-This script has been renamed scil_tractogram_uniformize_endpoints.py. Please
+This script has been renamed scil_bundle_uniformize_endpoints.py. Please
 change your existing pipelines accordingly.
 """
