+++ scripts/legacy/scil_validate_and_correct_bvecs.py
