+++ scripts/legacy/scil_validate_and_correct_eddy_gradients.py
