+++ scripts/legacy/scil_validate_bids.py
