+++ scripts/legacy/scil_verify_space_attributes_compatibility.py
