+++ scripts/scil_bundle_uniformize_endpoints.py
